move style loader and dependencies to bundler pkg
