+++ packages/docusaurus-bundler/package.json
@@ -31,16 +31,24 @@
     "@docusaurus/logger": "3.5.2",
     "@docusaurus/types": "3.5.2",
     "@docusaurus/utils": "3.5.2",
+    "autoprefixer": "^10.4.14",
     "babel-loader": "^9.1.3",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "clean-css": "^5.3.2",
     "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.1",
     "css-minimizer-webpack-plugin": "^5.0.1",
+    "cssnano": "^6.1.2",
+    "postcss": "^8.4.26",
+    "postcss-loader": "^7.3.3",
+    "file-loader": "^6.2.0",
     "fs-extra": "^11.1.1",
     "mini-css-extract-plugin": "^2.9.1",
+    "null-loader": "^4.0.1",
     "react-dev-utils": "^12.0.1",
     "terser-webpack-plugin": "^5.3.9",
     "tslib": "^2.6.0",
+    "url-loader": "^4.1.1",
     "webpack": "^5.88.1",
     "webpackbar": "^6.0.1"
   },
