+++ packages/docusaurus-bundler/src/index.ts
@@ -21,3 +21,7 @@ export {
   getCustomBabelConfigFilePath,
   createJsLoaderFactory,
 } from './loaders/jsLoader';
+
+export {createStyleLoadersFactory} from './loaders/styleLoader';
+
+export {default as babelPreset} from './loaders/babelPreset';
