+++ packages/docusaurus/package.json
@@ -42,7 +42,6 @@
     "@docusaurus/utils": "3.5.2",
     "@docusaurus/utils-common": "3.5.2",
     "@docusaurus/utils-validation": "3.5.2",
-    "autoprefixer": "^10.4.14",
     "boxen": "^6.2.1",
     "chalk": "^4.1.2",
     "chokidar": "^3.5.3",
@@ -50,25 +49,18 @@
     "combine-promises": "^1.1.0",
     "commander": "^5.1.0",
     "core-js": "^3.31.1",
-    "css-loader": "^6.8.1",
-    "css-minimizer-webpack-plugin": "^5.0.1",
-    "cssnano": "^6.1.2",
     "del": "^6.1.1",
     "detect-port": "^1.5.1",
     "escape-html": "^1.0.3",
     "eta": "^2.2.0",
     "eval": "^0.1.8",
-    "file-loader": "^6.2.0",
     "fs-extra": "^11.1.1",
     "html-minifier-terser": "^7.2.0",
     "html-tags": "^3.3.1",
     "html-webpack-plugin": "^5.5.3",
     "leven": "^3.1.0",
     "lodash": "^4.17.21",
-    "null-loader": "^4.0.1",
     "p-map": "^4.0.0",
-    "postcss": "^8.4.26",
-    "postcss-loader": "^7.3.3",
     "prompts": "^2.4.2",
     "react-dev-utils": "^12.0.1",
     "react-helmet-async": "^1.3.0",
@@ -83,7 +75,6 @@
     "shelljs": "^0.8.5",
     "tslib": "^2.6.0",
     "update-notifier": "^6.0.2",
-    "url-loader": "^4.1.1",
     "webpack": "^5.88.1",
     "webpack-bundle-analyzer": "^4.9.0",
     "webpack-dev-server": "^4.15.1",
