+++ packages/docusaurus/src/babel/preset.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+// TODO Docusaurus v4, do breaking change and expose babel preset cleanly
+export {babelPreset as default} from '@docusaurus/bundler';
